Consolidate reference-only components into reference/ (DEV-2338)
Move components whose only consumers are the reference app from
shared/common into reference/components: List, ListItem, LabelIcon,
QueryButton, SidebarItem, SidebarLayout, and Sidebar.module.css.

Verified via dependency-graph analysis + repo-wide greps (incl. e2e):
every importer of these files lives under frontend/src/metabase/reference.
Other similarly-named components (metabase/ui List, ListView viz,
AccordionList, ListField) are unrelated.

Shrinks shared/common and removes cross-module edges ahead of further
modularization.
