+++ frontend/src/metabase/reference/components/EditableReferenceHeader.tsx
@@ -3,8 +3,8 @@ import { memo } from "react";
 import { Link } from "react-router";
 import { t } from "ttag";
 
-import L from "metabase/common/components/List/List.module.css";
 import CS from "metabase/css/core/index.css";
+import L from "metabase/reference/components/List/List.module.css";
 import { Button, Ellipsified, Icon, TextInputBlurChange } from "metabase/ui";
 import type { IconName } from "metabase-types/api";
