+++ frontend/src/metabase/reference/components/Field.tsx
@@ -5,7 +5,6 @@ import { Link } from "react-router";
 import { P, match } from "ts-pattern";
 import { t } from "ttag";
 
-import S from "metabase/common/components/List/List.module.css";
 import { FIELD_SEMANTIC_TYPES_MAP } from "metabase/common/utils/fields";
 import CS from "metabase/css/core/index.css";
 import {
@@ -13,6 +12,7 @@ import {
   SemanticTypePicker,
 } from "metabase/metadata/components";
 import { getFieldCurrency } from "metabase/metadata/utils/field";
+import S from "metabase/reference/components/List/List.module.css";
 import { Box, Icon } from "metabase/ui";
 import { isTypeCurrency, isTypeFK } from "metabase-lib/v1/types/utils/isa";
 import type {
