+++ frontend/src/metabase/reference/components/LabelIcon/LabelIcon.module.css
