+++ frontend/src/metabase/reference/components/LabelIcon/LabelIcon.tsx
