+++ frontend/src/metabase/reference/components/LabelIcon/index.tsx
