+++ frontend/src/metabase/reference/components/QueryButton/index.tsx
