+++ frontend/src/metabase/reference/components/ReferenceHeader.tsx
@@ -3,9 +3,9 @@ import { memo } from "react";
 import { Link } from "react-router";
 import { t } from "ttag";
 
-import L from "metabase/common/components/List/List.module.css";
 import ButtonsS from "metabase/css/components/buttons.module.css";
 import CS from "metabase/css/core/index.css";
+import L from "metabase/reference/components/List/List.module.css";
 import { Ellipsified, Icon } from "metabase/ui";
 import type { IconName } from "metabase-types/api";
