+++ frontend/src/metabase/reference/components/Sidebar.module.css
