+++ frontend/src/metabase/reference/components/SidebarItem/SidebarItem.tsx
