+++ frontend/src/metabase/reference/components/SidebarItem/index.tsx
