+++ frontend/src/metabase/reference/components/SidebarLayout/SidebarLayout.tsx
