+++ frontend/src/metabase/reference/databases/FieldList.tsx
@@ -3,7 +3,6 @@ import { useFormik } from "formik";
 import { t } from "ttag";
 
 import { EmptyState } from "metabase/common/components/EmptyState";
-import S from "metabase/common/components/List/List.module.css";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
 import CS from "metabase/css/core/index.css";
 import { connect } from "metabase/redux";
@@ -13,6 +12,7 @@ import { EditHeader } from "metabase/reference/components/EditHeader";
 import EditableReferenceHeader from "metabase/reference/components/EditableReferenceHeader";
 import Field from "metabase/reference/components/Field";
 import F from "metabase/reference/components/Field.module.css";
+import S from "metabase/reference/components/List/List.module.css";
 import * as actions from "metabase/reference/reference";
 import { getIconForField } from "metabase-lib/v1/metadata/utils/fields";
 import type {
