+++ frontend/src/metabase/reference/segments/SegmentDetailContainer.tsx
@@ -2,10 +2,10 @@ import cx from "classnames";
 import type { Location } from "history";
 import { Component } from "react";
 
-import { SidebarLayout } from "metabase/common/components/SidebarLayout";
 import CS from "metabase/css/core/index.css";
 import { connect } from "metabase/redux";
 import * as metadataActions from "metabase/redux/metadata";
+import { SidebarLayout } from "metabase/reference/components/SidebarLayout";
 import * as actions from "metabase/reference/reference";
 import SegmentDetail from "metabase/reference/segments/SegmentDetail";
 import type { User } from "metabase-types/api";
