+++ frontend/src/metabase/reference/segments/SegmentSidebar.tsx
@@ -3,9 +3,9 @@ import { memo } from "react";
 import { t } from "ttag";
 
 import { Breadcrumbs } from "metabase/common/components/Breadcrumbs";
-import S from "metabase/common/components/Sidebar.module.css";
-import { SidebarItem } from "metabase/common/components/SidebarItem";
 import CS from "metabase/css/core/index.css";
+import S from "metabase/reference/components/Sidebar.module.css";
+import { SidebarItem } from "metabase/reference/components/SidebarItem";
 import MetabaseSettings from "metabase/utils/settings";
 import type { User } from "metabase-types/api";
